[fix] Build for windows#1

+++ libnet2/pool_node_data.cpp
@@ -17,7 +17,7 @@ PoolNodeData* PoolNodeData::set_coind_node(std::shared_ptr<CoindNodeData> _coind
 std::vector<ShareType> PoolNodeData::handle_get_shares(std::vector<uint256> hashes, uint64_t parents, std::vector<uint256> stops,
 													   addr_type peer_addr)
 {
-	parents = std::min(parents, 1000/hashes.size());
+	parents = std::min(parents, (uint64_t)1000/hashes.size());
 	std::vector<ShareType> shares;
 	for (auto share_hash : hashes)
 	{
